kitchen_display.py와 kitchen_node 병합, 인터페이스에서 detail_id 등 변수길이 길게 조정

+++ install/pizza_order_msgs/include/pizza_order_msgs/pizza_order_msgs/msg/detail/order_info__struct.h
@@ -24,8 +24,8 @@ extern "C"
 /// Struct defined in msg/OrderInfo in the package pizza_order_msgs.
 typedef struct pizza_order_msgs__msg__OrderInfo
 {
-  int8_t order_id;
-  int8_t table_number;
+  int16_t order_id;
+  int16_t table_number;
   rosidl_runtime_c__String status;
 } pizza_order_msgs__msg__OrderInfo;
